requirements.txt
Flask
Flasgger
python-dotenv
pymongo
flask_restful
requests
